+++ Implementation/AssistU/app/controllers/ProjectData.java
@@ -18,32 +18,42 @@ import static play.libs.Json.toJson;
 public class ProjectData extends Controller {
 
     public static Result project(Long pid) {
-        Person person = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(person != null) {
+
+        session().put("callback", routes.ProjectData.project(pid).absoluteURL(request()));
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
             Project.updateLastAccessed(pid);
             Project p = Project.find.byId(pid);
-            return ok(project.render("AssistU - Projects", person, p));
-        }else
+            return ok(project.render("AssistU - Projects", user, p));
+        } else {
+            session().put("callback", routes.ProjectData.project(pid).absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     static Form<Project> projectForm = Form.form(Project.class);
 
     public static Result createProjectPage() {
-        Person person = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(person != null)
-            return ok(projectNew.render("Create a new Project", projectForm, false, "", "", person));
-        else
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null)
+            return ok(projectNew.render("Create a new Project", projectForm, false, "", "", user));
+        else {
+            session().put("callback", routes.ProjectData.createProjectPage().absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     public static Result editProjectPage(Long pid) {
-        Person person = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
         Project p = Project.find.byId(pid);
-        if(person != null)
-            return ok(projectEdit.render("Edit Project " + p.name, p, projectForm, false, "", "", person));
-        else
+
+        if(user != null)
+            return ok(projectEdit.render("Edit Project " + p.name, p, projectForm, false, "", "", user));
+        else {
+            session().put("callback", routes.ProjectData.editProjectPage(pid).absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     /**
@@ -51,29 +61,40 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result createProject() {
-        Person person = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
         Form<Project> filledProjectForm = projectForm.bindFromRequest();
-        if(filledProjectForm.hasErrors()) {
-            return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
-                    "The input did not fulfill the requirements, please review your information", person));
-        } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
-                && Application.allowedTitleRegex(filledProjectForm.get().name))) {
-            return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
-                    "The input did not have the allowed format, please review your information", person));
-        } else {
-            Project projectData = filledProjectForm.get();
-            Project p = Project.create(projectData.folder, projectData.name, projectData.description, projectData.template);
-            Project.inviteOwner(p.id, person.id);
-            Role r = Role.find.where().eq("project",p).eq("person", person).findUnique();
-            r.accepted=true;
-            r.dateJoined=new Date();
-            r.update();
-            if(!p.template.equals("None")){
-                Event.defaultPlanningArticle(person, p);
-                p.planning=true;
-                p.save();}
-            return redirect(routes.ProjectData.project(p.id));
+
+
+        if (user != null) {
+            if (filledProjectForm.hasErrors()) {
+                return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
+                        "The input did not fulfill the requirements, please review your information", user));
+            } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
+                    && Application.allowedTitleRegex(filledProjectForm.get().name))) {
+                return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
+                        "The input did not have the allowed format, please review your information", user));
+            } else {
+                Project projectData = filledProjectForm.get();
+                Project p = Project.create(projectData.folder, projectData.name, projectData.description, projectData.template);
+                Project.inviteOwner(p.id, user.id);
+                Role r = Role.find.where().eq("project", p).eq("person", user).findUnique();
+                r.accepted = true;
+                r.dateJoined = new Date();
+                r.update();
+                if (!p.template.equals("None")) {
+                    Event.defaultPlanningArticle(user, p);
+                    p.planning = true;
+                    p.save();
+                }
+                Emailer.sendNotifyEmail("[assistU] New project "+p.name+ "Created",user,views.html.email.projectCreated.render(user,p));
+
+                return redirect(routes.ProjectData.project(p.id));
+            }
+
         }
+        //User did not have a session
+        session().put("callback", routes.ProjectData.createProject().absoluteURL(request()));
+        return Authentication.login();
     }
 
     /**
@@ -82,23 +103,30 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result editProject(Long pid) {
-        Person person = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(person)) {
-            Project p = Project.find.byId(pid);
-            Form<Project> filledProjectForm = projectForm.bindFromRequest();
-            if (filledProjectForm.hasErrors()) {
-                Logger.debug(filledProjectForm.errors().toString());
-                return badRequest(projectEdit.render("Something went wrong", p, filledProjectForm, true, "danger",
-                        "The input did not fulfill the requirements, please review your information", person));
-            } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
-                    && Application.allowedTitleRegex(filledProjectForm.get().name))) {
-                return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
-                        "The input did not have the allowed format, please review your information", person));
-            } else {
-                Project.edit(pid, filledProjectForm.get().folder, filledProjectForm.get().name, filledProjectForm.get().description);
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if (user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                Project p = Project.find.byId(pid);
+                Form<Project> filledProjectForm = projectForm.bindFromRequest();
+                if (filledProjectForm.hasErrors()) {
+                    Logger.debug(filledProjectForm.errors().toString());
+                    return badRequest(projectEdit.render("Something went wrong", p, filledProjectForm, true, "danger",
+                            "The input did not fulfill the requirements, please review your information", user));
+                } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
+                        && Application.allowedTitleRegex(filledProjectForm.get().name))) {
+                    return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
+                            "The input did not have the allowed format, please review your information", user));
+                } else {
+                    Project.edit(pid, filledProjectForm.get().folder, filledProjectForm.get().name, filledProjectForm.get().description);
+                }
             }
+            return redirect(routes.ProjectData.project(pid));
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.editProject(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.ProjectData.project(pid));
     }
 
     /**
@@ -109,11 +137,18 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result archiveProject(Long pid) {
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(u) && ProjectData.findAllOwners(pid).size() == 1) {
-            Project.archive(pid);
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user) && ProjectData.findAllOwners(pid).size() == 1) {
+                Project.archive(pid);
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.archiveProject(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     /*TODO SOHEIL: When a member gets invited, send him a message
@@ -123,34 +158,41 @@ public class ProjectData extends Controller {
      * @param pid: The project to which the user in the form has to be assigned
      * @return
      */
-    public static Result inviteMemberToProjectAs(Long pid){
+    public static Result inviteMemberToProjectAs(Long pid) {
         DynamicForm emailform = Form.form().bindFromRequest();
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        //See if the user requesting to invite someone is an Owner
-        if(ProjectData.findAllOwners(pid).contains(u)) {
-            Person person = Person.find.where().eq("email", emailform.get("email")).findUnique();
-            //See if the user that needs to be invited exists in the system
-            if(person != null) {
-                Project p = Project.find.byId(pid);
-                //There can not be a role relation between the invited user and the project,
-                // as it would be the user is already a member
-                if (Role.find.where().eq("project", p).eq("person", person).findUnique() == null) {
-                    //Pattern match the correct role for invitation
-                    if (emailform.get("role").equals("Owner")) {
-                        p.inviteOwner(p.id, person.id);
-                        Event.defaultPlanningArticle(person, p);
-                        p.planning = true;
-                        //            p.update();
-                        p.save();
-                    } else if (emailform.get("role").equals("Reviewer")) {
-                        p.inviteReviewer(p.id, person.id);
-                    } else {
-                        p.inviteGuest(p.id, person.id);
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if (user != null){
+            //See if the user requesting to invite someone is an Owner
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                Person user_invited = Person.find.where().eq("email", emailform.get("email")).findUnique();
+                //See if the user that needs to be invited exists in the system
+                if (user_invited != null) {
+                    Project p = Project.find.byId(pid);
+                    //There can not be a role relation between the invited user and the project,
+                    // as it would be the user is already a member
+                    if (Role.find.where().eq("project", p).eq("person", user_invited).findUnique() == null) {
+                        //Pattern match the correct role for invitation
+                        if (emailform.get("role").equals("Owner")) {
+                            Project.inviteOwner(p.id, user_invited.id);
+                            Event.defaultPlanningArticle(user, p);
+                            p.planning = true;
+                            //            p.update();
+//                            p.save();
+                        } else if (emailform.get("role").equals("Reviewer")) {
+                            Project.inviteReviewer(p.id, user_invited.id);
+                        } else {
+                            Project.inviteGuest(p.id, user_invited.id);
+                        }
                     }
                 }
             }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.inviteMemberToProjectAs(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     /*TODO SOHEIL: This function can send a message to all owners that a new user has joined the project
@@ -158,25 +200,45 @@ public class ProjectData extends Controller {
      */
     public static Result hasAccepted(Long pid){
         Project p = Project.find.byId(pid);
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Role r = Role.find.where().eq("person", u).eq("project", p).findUnique();
-        //See if there actually exists a (pending) role between the accepting user and project
-        if(r != null){
-            r.accepted=true;
-            r.dateJoined = new Date();
-            r.save();
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if( user != null) {
+            Role r = Role.find.where().eq("person", user).eq("project", p).findUnique();
+            //See if there actually exists a (pending) role between the accepting user and project
+            if (r != null) {
+                r.accepted = true;
+                r.dateJoined = new Date();
+                r.save();
+                List<Person> owners=findAllOwners(p.id);
+                owners.stream().forEach((u) -> {
+                    if(!u.equals(user))
+                    Emailer.sendNotifyEmail("[Assistu]"+ user.name + "has joined the project" , u ,views.html.email.project_joined.render(u,user,p) );
+                });
+            }
+            return redirect(routes.ProjectData.project(p.id));
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.hasAccepted(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.ProjectData.project(p.id));
     }
 
     public static Result hasDeclined(Long pid){
         Project p = Project.find.byId(pid);
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Role r = Role.find.where().eq("person", u).eq("project", p).findUnique();
-        if(r != null) {
-            r.delete();
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            Role r = Role.find.where().eq("person", user).eq("project", p).findUnique();
+            if (r != null) {
+                r.delete();
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.hasDeclined(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
+
     }
 
     /**
@@ -188,17 +250,31 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result removeMemberFromProject(Long uid, Long pid){
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(u)) {
-            Project.removeMemberFrom(pid, uid);
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                Project.removeMemberFrom(pid, uid);
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.removeMemberFromProject(uid,pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     public static Result leaveProject(Long pid){
-        Person u = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Project.removeMemberFrom(pid, u.id);
-        return redirect(routes.Application.project());
+
+        Person user = Person.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            Project.removeMemberFrom(pid, user.id);
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.leaveProject(pid).absoluteURL(request()));
+            return Authentication.login();
+        }
     }
 
     public static Result getProjectIdsAsJson(){
@@ -231,6 +307,7 @@ public class ProjectData extends Controller {
     public static Result getReviewerIdsAsJson(){
         List<Project> projects = PersonData.findActiveProjects();
         List<Long> result = new ArrayList<Long>();
+
         for(int i =0 ; i < projects.size(); i++){
             List<Person> projectreviewers = ProjectData.findAllReviewers(projects.get(i).id);
             for(int j = 0; j < projectreviewers.size(); j++){
@@ -267,6 +344,12 @@ public class ProjectData extends Controller {
         return ok(toJson(project));
     }
 
+
+    /**
+     * TODO : comment please
+     * @param pid
+     * @return
+     */
     public static List<Person> findAllAffiliatedUsers(Long pid){
         Project p = Project.find.byId(pid);
         List<Role> roles = Role.find.where().eq("project", p).findList();
@@ -277,6 +360,12 @@ public class ProjectData extends Controller {
         return persons;
     }
 
+
+    /**
+     * TODO : comment please
+     * @param pid
+     * @return
+     */
     public static List<Person> findAllOwners(Long pid){
         Project p = Project.find.byId(pid);
         List<Role> roles = Role.find.where().eq("project", p).eq("role", Role.OWNER).findList();
@@ -287,6 +376,12 @@ public class ProjectData extends Controller {
         return persons;
     }
 
+
+    /**
+     * TODO : comment please
+     * @param pid
+     * @return
+     */
     public static List<Person> findAllReviewers(Long pid){
         Project p = Project.find.byId(pid);
         List<Role> roles = Role.find.where().eq("project", p).eq("role", Role.REVIEWER).findList();
